pyproject.toml
[project]
name = "youtube-video-summarizer"
version = "0.1.0"
description = ""
authors = [
    {name = "RichardFFrank",email = "71610664+RichardFFrank@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "langchain (>=0.3.14,<0.4.0)",
    "tiktoken (>=0.8.0,<0.9.0)",
    "openai (>=1.59.8,<2.0.0)",
    "youtube-transcript-api (>=0.6.3,<0.7.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-groq (>=0.2.3,<0.3.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "aws-cdk-lib (>=2.176.0,<3.0.0)",
    "constructs (>=10.4.2,<11.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
